requirements.txt
# Core RL / ML
torch
gymnasium
numpy
pandas

# Exchange API
python-binance
python-dotenv

# Optional plotting / analysis
matplotlib
seaborn
tqdm
